chore: revert src20 to stamping; add progress to mint; update deploy content;
