+++ islands/stamping/src20/transfer/TransferContent.tsx
@@ -1,7 +1,7 @@
-import { FeeEstimation } from "../FeeEstimation.tsx";
+import { FeeEstimation } from "../../FeeEstimation.tsx";
 import { useSRC20Form } from "$islands/hooks/useSRC20Form.ts";
 
-export function MintContent(
+export function TransferContent(
   { trxType = "multisig" }: { trxType?: "olga" | "multisig" },
 ) {
   const {
@@ -16,7 +16,7 @@ export function MintContent(
     submissionMessage,
     walletError,
     apiError,
-  } = useSRC20Form("mint", trxType);
+  } = useSRC20Form("transfer", trxType);
 
   if (isLoading) {
     return <div>Loading...</div>;
@@ -26,33 +26,43 @@ export function MintContent(
     return <div>Error: Failed to load configuration</div>;
   }
 
-  const handleMintSubmit = async () => {
+  const handleTransferSubmit = async () => {
     try {
       await handleSubmit();
     } catch (error) {
-      console.error("Minting error:", error);
+      console.error("Transfer error:", error);
     }
   };
 
   return (
     <div class="flex flex-col w-full items-center gap-8">
-      <p class="bg-clip-text text-transparent bg-gradient-to-r from-[#440066] via-[#660099] to-[#8800CC] text-3xl md:text-6xl font-black mt-6 w-full text-center">
-        MINT SRC-20
+      <p class="text-[#5503A6] text-3xl md:text-6xl font-black mt-6 w-full text-center">
+        TRANSFER
       </p>
 
-      <div className="bg-gradient-to-br from-[#1F002E00] via-[#14001F7F] to-[#1F002EFF] p-6 w-full">
-        <div className="flex flex-col gap-6 w-full">
+      <div className="flex flex-col gap-6 bg-gradient-to-br from-[#1F002E00] via-[#14001F7F] to-[#1F002EFF] p-6 w-full">
+        <div class="w-full">
+          <input
+            type="text"
+            class="p-4 bg-[#999999] text-[#333333] placeholder:text-[#333333] font-medium w-full outline-none rounded-md focus:bg-[#CCCCCC]"
+            placeholder="Recipient address"
+            value={formState.toAddress}
+            onChange={(e) => handleInputChange(e, "toAddress")}
+          />
+          {formState.toAddressError && (
+            <p class="text-red-500 mt-2">{formState.toAddressError}</p>
+          )}
+        </div>
+
+        <div className="w-full flex gap-6">
           <div class="w-full">
             <input
               type="text"
-              class="p-4 bg-[#999999] text-[#333333] placeholder:text-[#333333] font-medium w-full outline-none rounded-md"
+              class="p-4 bg-[#999999] text-[#333333] placeholder:text-[#333333] font-medium w-full outline-none rounded-md focus:bg-[#CCCCCC]"
               placeholder="Token"
               value={formState.token}
               onChange={(e) => handleInputChange(e, "token")}
             />
-            {formState.tokenError && (
-              <p class="text-red-500 mt-2">{formState.tokenError}</p>
-            )}
           </div>
 
           <div class="w-full">
@@ -60,14 +70,11 @@ export function MintContent(
               type="text"
               inputMode="numeric"
               pattern="[0-9]*"
-              class="p-4 bg-[#999999] text-[#333333] placeholder:text-[#333333] font-medium w-full outline-none rounded-md"
+              class="p-4 bg-[#999999] text-[#333333] placeholder:text-[#333333] font-medium w-full outline-none rounded-md  focus:bg-[#CCCCCC]"
               placeholder="Amount"
               value={formState.amt}
               onChange={(e) => handleInputChange(e, "amt")}
             />
-            {formState.amtError && (
-              <p class="text-red-500 mt-2">{formState.amtError}</p>
-            )}
           </div>
         </div>
       </div>
@@ -76,13 +83,13 @@ export function MintContent(
         <FeeEstimation
           fee={formState.fee}
           handleChangeFee={handleChangeFee}
-          type="src20-mint"
+          type="src20-transfer"
           fileType="application/json"
           fileSize={formState.jsonSize}
           BTCPrice={formState.BTCPrice}
           onRefresh={fetchFees}
           isSubmitting={isSubmitting}
-          onSubmit={handleMintSubmit}
+          onSubmit={handleTransferSubmit}
         />
 
         {apiError && (
